Implement search && fix several display issues
* Implement search functionality using Google's "Custom Search"
product; google_search.js, search_results.jade are the support
files for this feature.

* Introduce initial sitemap.txt for site; this allows more efficient
indexing to be performed by search engines.

* layout.jade: Add anchors to 'office location' && 'contact' links.

* contact.jade: Fix 'Send Message' button by changing it from an input
element to a button.
* contact.jade: Implement caution notice as an unordered-list;
'contact_form ul' && 'contact_form li' CSS styles are responsible
for this.

* style.css: Change blue_button from a class to an ID selector.
* style.css: Remove .header-n class selectors.
* style.css: Rename page_wrap, content_wrap to page_container,
content_container.
* style.css: Introduce navi_container; this along with couple other
changes to nav-related selectors has resolved the alignment issues
that has been plaguing me; we can now align the navigational links
left, center or right and have the search input stay to the right
of us. (In short, turning the form elements to display inline
instead of block is the fix).

* Remove css, img, js symbolic links; this was causing duplicate search
entries to appear...

* Remove public/images/caution.png; it has been encoded as base64
within style.css.