Preliminary server-side validation for contact page
...via express-validator package; said page now sends HTTP POST upon
submission.

* contact.jade: Use hidden labels for un-named fields (phone
extension, last name, etc.); for sake of HTML validation...

* contact.jade: Use HTML5 placeholder & required fields & remove JS
placeholder implementation; dealing with the default values on the
server-side prove more of a headache than it is worth!

* Introduce site_message.jade template; this can be used for display
of server-side messages.

* Introduce 'offline' mode so that we can continue working on the
site through net outages -- such as this very moment -_-.

* style.css: Relocate searchbar selector ID to google_cse.css;
this lets us optionally disable it when in 'offline' mode.

* Use HTML5 input types for all form elements.

* Remove users.js; this was a template (default generated) file.