Introduce web form to email functionality with SendGrid
Mock email objects in the development environment, and the real deal
in testing and production environments.

By default, both plain-text and HTML email copies are sent,
leaving it up to the email client which to prefer displaying.

* Introduce 'testing' environment option.

* contact_success.jade: Relocate the tabular display of contact form
fields to views/email_html_template.jade; this is to support HTML
emails.

* Introduce views/email_text_template.jade; this is to support
plain-text e-mails.

* README.md: Add some documentation for what happens in testing
and production environments, and include info on how to set the
SENDGRID user name and password environment variables.