Introduce welcome banner for home page
* Add minimal contact form elements to home page; as per
client's request. Logic for the form has not yet been
implemented.
  Add jQuery.SimpleModal JS library for modal popup windows;
this is used for displaying the email disclaimer for the form;
this maximizes vertical space.

* Remove disclaimer.md and use disclaimer.jade as an
include instead, re-purposed as the e-mail form
disclaimer.

* Refactor base font size to 16px; this should hopefully be more
deterministic; previously, we were assuming font sizes based on
browser defaults (which will vary from system to system...).

* site_footer.jade: Refactor to use two columns.

* layout.jade: Remove maximum-scale, user-scalable meta blocks.

* layout.jade: Add minimal-ui meta block; this tells Mobile Safari
to not include the upper and lower UI elements of the browser upon
page load, maximizing vertical space.

* Reduce the maximal length of the contact form's message field
to 4096 (from 8192).